chore: 패키지명 오류 수정
