--- delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/applicatoin/dto/CreateDeliveryResponse.java
@@ -1,36 +1,0 @@
-package com.eighteenthstreet.deliveryservice.applicatoin.dto;
-
-import java.util.UUID;
-
-import com.eighteenthstreet.deliveryservice.domain.model.Delivery;
-import com.eighteenthstreet.deliveryservice.domain.model.DeliveryStatus;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-@Builder
-public class CreateDeliveryResponse {
-	private UUID deliveryId;
-	private UUID orderId;
-	private UUID startHubId;
-	private UUID endHubId;
-	private DeliveryStatus status;
-
-	public static CreateDeliveryResponse fromEntity(Delivery delivery) {
-		return CreateDeliveryResponse.builder()
-			.deliveryId(delivery.getDeliveryId())
-			.orderId(delivery.getOrderId())
-			.startHubId(delivery.getStartHubId())
-			.endHubId(delivery.getEndHubId())
-			.status(delivery.getStatus())
-			.build();
-	}
-
-}
